modify SnapPermissions to take the permissions as a prop
