+++ app/components/Views/Snaps/components/SnapPermissions/test/SnapPermission.test.tsx
@@ -6,14 +6,13 @@ import {
   SNAP_PERMISSION_CELL,
 } from '../../../../../../constants/test-ids';
 import { SnapPermissions as SnapPermissionsType } from '@metamask/snaps-utils';
-import renderWithProvider from '../../../../../../util/test/renderWithProvider';
 import {
   SubjectPermissions,
   PermissionConstraint,
 } from '@metamask/permission-controller';
+import { render } from '@testing-library/react-native';
 
 describe('SnapPermissions', () => {
-  const mockSnapId = '@metamask/mock-snap';
   const mockDate = 1684964145490;
   const mockDate2 = 1686081721987;
   const longRunningTitle = 'Run indefinitely';
@@ -175,22 +174,6 @@ describe('SnapPermissions', () => {
 
   beforeEach(() => {
     jest.clearAllMocks();
-  });
-
-  const mockEngineState = (
-    mockPermissions: SubjectPermissions<PermissionConstraint>,
-  ) => ({
-    engine: {
-      backgroundState: {
-        PermissionController: {
-          subjects: {
-            '@metamask/mock-snap': {
-              permissions: mockPermissions,
-            },
-          },
-        },
-      },
-    },
   });
 
   it('renders permissions correctly', () => {
@@ -360,24 +343,10 @@ describe('SnapPermissions', () => {
       },
     };
 
-    const engineState = {
-      engine: {
-        backgroundState: {
-          PermissionController: {
-            subjects: {
-              '@metamask/mock-snap': {
-                permissions: mockPermissions,
-              },
-            },
-          },
-        },
-      },
-    };
-
-    const { getAllByTestId } = renderWithProvider(
-      <SnapPermissions snapId={mockSnapId} />,
-      { state: engineState },
+    const { getAllByTestId } = render(
+      <SnapPermissions permissions={mockPermissions} />,
     );
+
     const permissionCells = getAllByTestId(SNAP_PERMISSION_CELL);
     const permissionCellTitles = getAllByTestId(SNAP_PERMISSIONS_TITLE);
     const permissionCellDates = getAllByTestId(SNAP_PERMISSIONS_DATE);
@@ -475,9 +444,8 @@ describe('SnapPermissions', () => {
       },
     };
 
-    const { getAllByTestId } = renderWithProvider(
-      <SnapPermissions snapId={mockSnapId} />,
-      { state: mockEngineState(mockPermissions) },
+    const { getAllByTestId } = render(
+      <SnapPermissions permissions={mockPermissions} />,
     );
     const permissionCellDates = getAllByTestId(SNAP_PERMISSIONS_DATE);
 
@@ -506,9 +474,8 @@ describe('SnapPermissions', () => {
         date: mockDate2,
       },
     };
-    const { getAllByTestId } = renderWithProvider(
-      <SnapPermissions snapId={mockSnapId} />,
-      { state: mockEngineState(mockPermissions) },
+    const { getAllByTestId } = render(
+      <SnapPermissions permissions={mockPermissions} />,
     );
     const permissionCells = getAllByTestId(SNAP_PERMISSIONS_TITLE);
 
@@ -535,9 +502,8 @@ describe('SnapPermissions', () => {
       },
     };
 
-    const { getAllByTestId } = renderWithProvider(
-      <SnapPermissions snapId={mockSnapId} />,
-      { state: mockEngineState(mockPermissions) },
+    const { getAllByTestId } = render(
+      <SnapPermissions permissions={mockPermissions} />,
     );
     const permissionCells = getAllByTestId(SNAP_PERMISSIONS_TITLE);
 
@@ -563,9 +529,8 @@ describe('SnapPermissions', () => {
       },
     };
 
-    const { getAllByTestId } = renderWithProvider(
-      <SnapPermissions snapId={mockSnapId} />,
-      { state: mockEngineState(mockPermissions) },
+    const { getAllByTestId } = render(
+      <SnapPermissions permissions={mockPermissions} />,
     );
     const permissionCells = getAllByTestId(SNAP_PERMISSIONS_TITLE);
 
@@ -592,9 +557,8 @@ describe('SnapPermissions', () => {
       },
     };
 
-    const { queryAllByTestId } = renderWithProvider(
-      <SnapPermissions snapId={mockSnapId} />,
-      { state: mockEngineState(mockPermissions) },
+    const { queryAllByTestId } = render(
+      <SnapPermissions permissions={mockPermissions} />,
     );
     const permissionCells = queryAllByTestId(SNAP_PERMISSIONS_TITLE);
 
@@ -617,9 +581,8 @@ describe('SnapPermissions', () => {
       },
     };
 
-    const { getAllByTestId } = renderWithProvider(
-      <SnapPermissions snapId={mockSnapId} />,
-      { state: mockEngineState(mockPermissions) },
+    const { getAllByTestId } = render(
+      <SnapPermissions permissions={mockPermissions} />,
     );
     const permissionCellTitles = getAllByTestId(SNAP_PERMISSIONS_TITLE);
 
@@ -792,9 +755,8 @@ describe('SnapPermissions', () => {
       },
     };
 
-    const { getAllByTestId } = renderWithProvider(
-      <SnapPermissions snapId={mockSnapId} />,
-      { state: mockEngineState(mockPermissions) },
+    const { getAllByTestId } = render(
+      <SnapPermissions permissions={mockPermissions} />,
     );
     const permissionCellTitles = getAllByTestId(SNAP_PERMISSIONS_TITLE);
 
@@ -886,9 +848,8 @@ describe('SnapPermissions', () => {
       },
     };
 
-    const { getAllByTestId } = renderWithProvider(
-      <SnapPermissions snapId={mockSnapId} />,
-      { state: mockEngineState(invalidMockPermissions) },
+    const { getAllByTestId } = render(
+      <SnapPermissions permissions={invalidMockPermissions} />,
     );
     const permissionCellTitles = getAllByTestId(SNAP_PERMISSIONS_TITLE);
 
@@ -927,9 +888,8 @@ describe('SnapPermissions', () => {
       },
     };
 
-    const { getAllByTestId } = renderWithProvider(
-      <SnapPermissions snapId={mockSnapId} />,
-      { state: mockEngineState(mockPermissions) },
+    const { getAllByTestId } = render(
+      <SnapPermissions permissions={mockPermissions} />,
     );
     const permissionCellTitles = getAllByTestId(SNAP_PERMISSIONS_TITLE);
 
@@ -959,9 +919,8 @@ describe('SnapPermissions', () => {
       },
     };
 
-    const { getAllByTestId } = renderWithProvider(
-      <SnapPermissions snapId={mockSnapId} />,
-      { state: mockEngineState(mockPermissions) },
+    const { getAllByTestId } = render(
+      <SnapPermissions permissions={mockPermissions} />,
     );
     const permissionCellTitles = getAllByTestId(SNAP_PERMISSIONS_TITLE);
 
@@ -1029,10 +988,7 @@ describe('SnapPermissions', () => {
   });
 
   it('renders correctly with no permissions', () => {
-    const { queryByTestId } = renderWithProvider(
-      <SnapPermissions snapId={mockSnapId} />,
-      { state: mockEngineState({}) },
-    );
+    const { queryByTestId } = render(<SnapPermissions permissions={{}} />);
     expect(queryByTestId(SNAP_PERMISSION_CELL)).toBeNull();
   });
 });
